Issue # 5920: Refactor linked lists
Move the linked list code out for reuse.
